packages/protocols/x402-aws-agentcore/payer-agent/pyproject.toml
[project]
name = "x402-payer-agent"
version = "0.1.0"
description = "AI agent for x402 payment decisions using Strands Agents and AgentCore"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT-0"}
authors = [
    {name = "x402 Demo"}
]
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "coinbase-agentkit>=0.1.0",
    "boto3>=1.35.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    # FastAPI for the web UI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # OpenTelemetry for distributed tracing
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-propagator-aws-xray>=1.0.0",
    "opentelemetry-sdk-extension-aws>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
